requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
pydantic==2.10.6
python-dotenv==1.0.1
pytest==8.3.5
httpx==0.28.1
openai==1.68.2
google-genai==1.7.0

